Update 5-16-17
- added framework for different weapon types + levels of each weapon

- refactored shoot script

- added shiny new particle effects for enemy/player death as well as
powerup get

- added new music tracks

- changed all audiosources to be slowed when in slow mo

- A BUNCH OF OTHER STUFF

+++ Assets/Scripts/Enemy.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Enemy : MonoBehaviour
 {
 
-    public int health = 0;
+    public float health = 0;
     public float moveSpeed = 0;
     public float attackSpeed = 0;
