+++ Assets/Scripts/EnemyModifier.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class EnemyModifier : MonoBehaviour {
 
-    public static int bonusHealth = 0;
+    public static float bonusHealth = 0;
     public static float bonusAttackSpeed = 0f;
     public static int bonusSpeed = 0;
     public static float bonusSpawnRate = 0f;
